[dev] pull pytest plugin out of core and into monorepo root
